[PSL-1180] pasteld: check & add missing insightexplorer code (#254)
* [PSL-1180] pasteld: check & add missing insightexplorer code

 pasteld v2.3.1-beta1
 - moved all insightexplorer code and adapted to pasteld:
    - mempool: add memory address & spent indices
	- mempool: remove address & spend indices from internal maps when transaction is removed from mempool
	- mempool: updated memory DynamicUsage for insight-related structures
	- ConnectBlock: collect info for insightexplorer indices from vin/vout transactions in a block,
	  add/update indices in txdb
    - DisconnectBlock: collect info for insightexplorer indices from transactions in a block and
	  erase from txdb
	- add/update/delete indices from txdb (address,spend and timestamp indices)
 - refactored checks for txdb/cmdline options mismatches for fTxIndex, fInsightExplorer flags
 - refactored serialization/deserialization for address,spent,timestamp indices storage in txdb
 - refactored logging for LevelDB create/open operations
 - removed static from constexpr declared outside classes/namespaces
 - moved experimental features initialization to the separate file, removed requirement for
   "experimentalfeatures" option to enable insightexplorer
 - added new python tests for insightexplorer: addressindex.py, spentindex.py, timestampindex.py
 - added new insightexplorer RPC APIs:
    - getblockdeltas blockhash
	  Returns the txid and index where an output is spent
	- getblockhashes high low ( {"noOrphans": true|false, "logicalTimes": true|false}
      Returns array of hashes of blocks within the timestamp range provided, greater or equal to low, less than high.
	- getaddressmempool {addresses: [taddr, ...]}
	  Returns all mempool deltas for an address.
	- getaddresstxids {"addresses": ["taddr", ...], ("start": n), ("end": n)}
	  Returns the transaction ids for given transparent addresses within the given (inclusive) block height range, default is the full blockchain.
	- getaddressbalance {"addresses": ["taddr", ...]}
	  Returns the balance for addresses.
	- getaddressdeltas {"addresses": ["taddr", ...], ("start": n), ("end": n), ("chainInfo": true|false)}
	  Returns all changes for an address. Returns information about all changes to the given transparent addresses within the given (inclusive) block height range, default is the full blockchain.
    - getaddressutxos {"addresses": ["taddr", ...], ("chainInfo": true|false)}
	  Returns all unspent outputs for an address.
    - getspentinfo {"txid": "txid", "index": n}
	  Returns the txid and index where an output is spent.
 - refactored getting/validating start/end height RPC parameters
 - added google tests for insightexplorer RPC APIs calls (validate parameters, API disabled messages)

* fixes

* fixes

* fixes

+++ src/script/scripttype.h
@@ -1,10 +1,12 @@
 #pragma once
+// Copyright (c) 2018-2024 The Pastel Core developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
 
 // insightexplorer, there may be more script types in the future
-typedef enum class _ScriptType : int
+typedef enum class _ScriptType : uint8_t
 {
     UNKNOWN = 0,
-    P2PKH = 1,
-    P2SH = 2,
+    P2PKH = 1, // pay to pubkey hash (OP_DUP OP_HASH160 <PubKeyHash> OP_EQUALVERIFY OP_CHECKSIG)
+    P2SH = 2,  // pay to script hash (OP_HASH160 <ScriptHash> OP_EQUAL)
 } ScriptType;
-
